게시글 목록 쿼리 부하 테스트 후 게시글 도메인 수정 테스트 1(서브쿼리) - 처리량 1138.6/s 테스트 2(조인) - 처리량 1584.2/s 테스트 3(게시글 도메인에 컬럼 추가) - 처리량 1703.2/s
테스트 1번은 성능이 좋지않고 테스트 2번과 3번에서 트레이드오프가 있다

+++ interview/src/main/java/inflearn/interview/post/controller/response/PostResponse.java
@@ -47,7 +47,7 @@ public class PostResponse implements BaseDTO {
     private boolean isLiked;
 
     @QueryProjection
-    public PostResponse(Long userId, String userName, Long postId, String postTitle, String content, String category, LocalDateTime time, LocalDateTime updateTime, int numOfLike, Long numOfComment, String tag, String image) {
+    public PostResponse(Long userId, String userName, Long postId, String postTitle, String content, String category, LocalDateTime time, LocalDateTime updateTime, int numOfLike, int numOfComment, String tag, String image) {
         this.userId = userId;
         this.userName = userName;
         this.postId = postId;
@@ -57,7 +57,7 @@ public class PostResponse implements BaseDTO {
         this.time = time;
         this.updateTime = updateTime;
         this.numOfLike = numOfLike;
-        this.numOfComment = Math.toIntExact(numOfComment);
+        this.numOfComment = numOfComment;
         if (tag != null) {
             this.tag = entityToDtoTagConverter(tag);
         }
